+++ .storybook/stories/DataGrid.stories.tsx
@@ -70,7 +70,7 @@ DecisionList.args = {
   children: DecisionListMock.map((gridItem) => (
     <DataGridRow
       checkbox={true}
-      checkboxTooltip='Označiť'
+      checkboxTooltip="Označiť"
       active={!gridItem.inactive}
       key={gridItem.id}
       moreOptions={[
@@ -79,7 +79,7 @@ DecisionList.args = {
         <p key={3}>Sprievodca schránkou</p>,
         <p key={4}>Zobraziť detail</p>
       ]}
-      moreOptionsTooltip='Možnosti'
+      moreOptionsTooltip="Možnosti"
     >
       <DataGridRowValue>
         <div>
@@ -225,6 +225,45 @@ WithoutHeader.args = {
     </DataGridRow>
   ))
 };
+
+export const WithClickableBody = Template.bind({});
+WithClickableBody.args = {
+  checkboxEverything: true,
+  children: DecisionListMock.map((gridItem) => (
+    <DataGridRow
+      checkbox={true}
+      active={!gridItem.inactive}
+      key={gridItem.id}
+      moreOptions={[
+        <p key={1}>Exportovať</p>,
+        <p key={2}>Archivovať</p>,
+        <p key={3}>Sprievodca schránkou</p>,
+        <p key={4}>Zobraziť detail</p>
+      ]}
+      moreOptionsTooltip="Možnosti"
+      buttonProps={{
+        onClick: () => console.log(`You clicked on ${gridItem.id}`)
+      }}
+    >
+      <DataGridRowValue>
+        <div>
+          <div className="flex gap-2.5 items-center min-h-[2.375rem]">
+            <div className={!gridItem.inactive ? 'font-bold' : ''}>{gridItem.title}</div>
+            {gridItem.titleTag}
+          </div>
+          {gridItem.text}
+        </div>
+      </DataGridRowValue>
+      <DataGridRowValue align="right" className="tb2:flex hidden flex-wrap justify-end gap-2.5">
+        <DataGridTags tags={gridItem.tags} />
+      </DataGridRowValue>
+      <DataGridRowValue align="right" className="tb2:min-w-[10.375rem] tb2:max-w-[10.375rem]">
+        {gridItem.date}
+      </DataGridRowValue>
+    </DataGridRow>
+  ))
+};
+
 export const DecisionSharingList = Template.bind({});
 DecisionSharingList.args = {
   headRow: (